Create index.html and images directory
This is the first phase of the 'landing page' project.  I have created
the index.html file and a subdirectory to hold images.  I have also laid
out the HTML frame of the page.

Next step will be to come up with an actual theme for the page, find
some free images, and start styling with CSS.