Complete the landing-page project
I noticed that when I loaded my page in Safari that the Roboto font was not
being used.  I researched and added support for loading the font through the
Google API and it now works.

I also exchanged the stock image in the hero section for a version of the
app logo.

This should complete the project and I will now make it live and proceed
to the JavaScript portion of The Odin Project.